requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astunparse==1.6.3
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
fastapi==0.136.1
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.3.0
gast==0.4.0
google-auth==2.49.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.10.1
icrawler==0.6.10
idna==3.11
Jinja2==3.1.6
keras==3.14.0
libclang==18.1.1
lxml==6.0.4
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.1.0
networkx==3.6.1
numpy==1.26.0
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==3.0.2
pillow==12.2.0
protobuf==4.25.9
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-multipart==0.0.26
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
rich==15.0.0
safetensors==0.7.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
starlette==1.0.0
sympy==1.14.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-estimator==2.13.0
tensorflow-intel==2.16.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
tokenizers==0.22.2
torch==2.11.0
torchaudio==2.11.0+cpu
torchvision==0.26.0
tqdm==4.67.3
transformers==5.5.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
Werkzeug==3.1.8
wrapt==2.1.2
